requirements.txt
# Flask and Extensions
flask>=3.0.0
flask-cors>=4.0.0
flask-login>=0.6.3
python-dotenv>=1.0.0
werkzeug>=3.0.0

# Document Processing
PyPDF2>=3.0.0
python-docx>=1.0.0
pandas>=2.0.0
openpyxl>=3.1.0

# AI and Vector Search
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
numpy>=1.24.0

# Utilities
pydantic>=2.0.0
requests>=2.31.0
uuid6>=2024.1.12
